+++ LojaSaoJoao/src/main/java/com/br/fontana/digital/service/UsuarioService.java
@@ -6,11 +6,39 @@ import com.br.fontana.digital.response.UsuarioResponse;
 
 public interface UsuarioService {
 
+	/**
+	 * 
+	 * @param userName
+	 * @return
+	 * @throws Exception
+	 */
 	public UsuarioResponse findByUsername(String userName) throws Exception;
 	
-	public UsuarioResponse findByCdUsuario(Integer cdUsuario) throws Exception;
+	/**
+	 * 
+	 * @param cdUsuario
+	 * @return
+	 * @throws Exception
+	 */
+	public UsuarioResponse findByCdUsuario(Long cdUsuario) throws Exception;
 	
+	/**
+	 * 
+	 * @param cdLoja
+	 * @return
+	 * @throws Exception
+	 */
 	public List<UsuarioResponse> findByLoja(Integer cdLoja) throws Exception;
 	
+	/**
+	 * 
+	 * @param cdLoja
+	 * @param cdGrau
+	 * @param cdSituacao
+	 * @param mes
+	 * @param cdTipo
+	 * @return
+	 * @throws Exception
+	 */
 	public List<UsuarioResponse> findByFilter(Integer cdLoja, Integer cdGrau, Integer cdSituacao, Integer mes,  Integer cdTipo) throws Exception;
 }
